+++ providers/index.js
@@ -1,15 +1,5 @@
-import * as bing from './bing'
-import * as esri from './esri'
-import * as openstreetmap from './openstreetmap'
-import * as strava from './strava'
-import * as digitalglobe from './digitalglobe'
-import * as mapbox from './mapbox'
-
-export {
-  digitalglobe,
-  mapbox,
-  esri,
-  bing,
-  openstreetmap,
-  strava
+module.exports = {
+  esri: require('./esri'),
+  bing: require('./bing'),
+  openstreetmap: require('./openstreetmap')
 }